fix(ci): order built-in plugin packaging
