+++ .github/workflows/test.yml
@@ -63,7 +63,7 @@ jobs:
             printf '%s\n' "Changed files:" "$changed_files"
             while IFS= read -r file; do
               case "$file" in
-                assets/*|packages/templates/*|package.json|pnpm-lock.yaml|pnpm-workspace.yaml|turbo.json|scripts/*|.github/workflows/test.yml)
+                assets/*|eval/*|packages/templates/*|doctor.config.jsonc|promptfooconfig.yaml|package.json|pnpm-lock.yaml|pnpm-workspace.yaml|turbo.json|scripts/*|.github/workflows/test.yml)
                   full=true
                   ;;
               esac
@@ -206,8 +206,20 @@ jobs:
           set -euo pipefail
           start=$SECONDS
           if [[ "$FULL_RUN" == "true" ]]; then pnpm turbo run build --summarize; else pnpm turbo run build --affected --summarize; fi
-          if [[ "$FULL_RUN" == "true" ]]; then pnpm turbo run test:packaging --filter=@sero/desktop --summarize; else pnpm turbo run test:packaging --filter=@sero/desktop --affected --summarize; fi
           echo "Build took $((SECONDS - start))s"
+      - name: Packaging tests
+        shell: bash
+        env:
+          FULL_RUN: ${{ needs.changes.outputs.full }}
+          TURBO_TEAM: ${{ secrets.TURBO_TEAM }}
+          TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
+          TURBO_SCM_BASE: ${{ needs.changes.outputs.base }}
+          TURBO_SCM_HEAD: ${{ needs.changes.outputs.head }}
+        run: |
+          set -euo pipefail
+          start=$SECONDS
+          if [[ "$FULL_RUN" == "true" ]]; then pnpm turbo run test:packaging --filter=@sero/desktop --summarize; else pnpm turbo run test:packaging --filter=@sero/desktop --affected --summarize; fi
+          echo "Packaging tests took $((SECONDS - start))s"
       - name: Upload Turbo run summaries
         if: always()
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7
