+++ turbo.json
@@ -5,6 +5,22 @@
       "dependsOn": ["^build"],
       "outputs": ["dist/**"]
     },
+    "@sero/desktop#build": {
+      "dependsOn": [
+        "^build",
+        "@sero-ai/plugin-admin#build",
+        "@sero-ai/plugin-cron#build",
+        "@sero-ai/plugin-design-library#build",
+        "@sero-ai/plugin-git#build",
+        "@sero-ai/plugin-graphify#build",
+        "@sero-ai/plugin-mcp#build",
+        "@sero-ai/plugin-orchestrator#build",
+        "@sero-ai/plugin-usage#build",
+        "@sero-ai/plugin-user-feedback#build",
+        "@sero-ai/plugin-web#build"
+      ],
+      "outputs": ["dist/**"]
+    },
     "dev": {
       "cache": false,
       "persistent": true
